now is tidi

+++ engine/src/toast/physics/narrow_phase.hpp
@@ -15,7 +15,7 @@
 
 namespace physics {
 
-enum class NarrowPhasePairType : size_t {
+enum class NarrowPhasePairType : uint8_t {
 	sphere_sphere,
 	sphere_box,
 	sphere_capsule,
