+++ engine/src/toast/world/workspace_events.hpp
@@ -266,7 +266,7 @@ struct EditorCameraLook : Event<EditorCameraLook> {
 	EditorCameraLook(float dx, float dy) : dx(dx), dy(dy) { }
 };
 
-enum class EditorCameraMode : uint32_t {
+enum class EditorCameraMode : uint8_t {
 	free = 0,
 	orbit = 1
 };
